fixed bug at statetime calculation of sheetanimation

+++ core/src/com/mygdx/game/SheetAnimation.java
@@ -91,7 +91,7 @@ public class SheetAnimation implements Disposable {
         if(_startTimestamp == -1){
             _startTimestamp = TimeUtils.millis();
         }
-        float sampleTime = (TimeUtils.millis() - _startTimestamp) / 1000;
+        float sampleTime = (float)(TimeUtils.millis() - _startTimestamp) / 1000;
         TextureRegion frame = _animation.getKeyFrame(sampleTime, true);
         batch.draw(frame, _position.X, _position.Y);
     }
